backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
opencv-python==4.12.0.88
numpy==2.2.6
pytesseract==0.3.13
python-multipart==0.0.20
pydantic==2.11.7
doxapy>=0.9.2
paddleocr>=3.0.0,<3.5.0
paddlepaddle>=3.0.0,<3.3.0
PyMuPDF>=1.26.0,<1.27.0
Pillow>=11.0.0,<12.0.0
ImageHash>=4.3.1,<5.0.0
